提交

+++ src/com/neuedu/basic2/demo2/HourlyEmployee.java
@@ -6,19 +6,19 @@ package com.neuedu.basic2.demo2;
  * @date 2020/2/6 0:29
  */
 public class HourlyEmployee extends ColaEmployee {
-    private int hourlywage;  // 每小时的工资
-    private int hourage;  // 每个月工作的小时数
+    private Integer hourlywage;  // 每小时的工资
+    private Integer hourage;  // 每个月工作的小时数
 
-    public HourlyEmployee(String name,int hourlywage,int hourage,int month) {
+    public HourlyEmployee(String name,Integer hourlywage,Integer hourage,int month) {
         super(name,month);
         this.hourlywage = hourlywage;
         this.hourage = hourage;
     }
 
     @Override
-    public double getSalary(int month) {
+    public Double getSalary(Integer month) {
         if( hourage <= 160){
-            return hourlywage * hourage;
+            return Double.valueOf(hourlywage * hourage);
         }else{
             return 160 * hourlywage + (hourage-160) * 1.5 * hourlywage;
         }
